requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==12.1.0
setuptools==80.9.0
tkinterdnd2==0.4.3
wheel==0.45.1
